replaced hard waits with explicit wait and updated the success message formats
