+++ src/test/java/io/github/mfaisalkhatri/pages/lambdatestecommerce/ConfirmOrderPage.java
@@ -3,12 +3,12 @@ package io.github.mfaisalkhatri.pages.lambdatestecommerce;
 import static io.github.mfaisalkhatri.drivers.DriverManager.getDriver;
 import static org.testng.Assert.assertEquals;
 
+import java.text.MessageFormat;
 import java.time.Duration;
 
 import io.github.mfaisalkhatri.data.BillingData;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.interactions.Actions;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
@@ -22,13 +22,12 @@ public class ConfirmOrderPage {
     }
 
     public OrderSuccessPage confirmOrder () {
-        final Actions actions = new Actions (getDriver ());
         confirmOrderBtn ().click ();
         return new OrderSuccessPage ();
     }
 
     public ConfirmOrderPage verifyPageHeader () {
-        final WebDriverWait wait = new WebDriverWait (getDriver (), Duration.ofSeconds (20));
+        final WebDriverWait wait = new WebDriverWait (getDriver (), Duration.ofSeconds (10));
         assertEquals (wait.until (ExpectedConditions.visibilityOfElementLocated (By.tagName ("h1")))
             .getText (), "Confirm Order");
         return this;
@@ -42,9 +41,11 @@ public class ConfirmOrderPage {
     }
 
     public ConfirmOrderPage verifyShippingAddress (final BillingData billingData) {
+        final String expectedMessage = "{0} {1}\n{2}\n{3} {4}\n{5},{6}";
         assertEquals (getDriver ().findElement (By.cssSelector ("#content > div.row > div:nth-child(2) > div > div"))
-                .getText (),
-            billingData.getFirstName () + " " + billingData.getLastName () + "\n" + billingData.getAddressLineOne () + "\n" + billingData.getCity () + " " + billingData.getPostCode () + "\n" + billingData.getState () + "," + billingData.getCountry ());
+            .getText (), MessageFormat.format (expectedMessage, billingData.getFirstName (), billingData.getLastName (),
+            billingData.getAddressLineOne (), billingData.getCity (), billingData.getPostCode (),
+            billingData.getState (), billingData.getCountry ()));
 
         return this;
     }
