+++ src/test/java/io/github/mfaisalkhatri/tests/saucedemo/SauceDemoTests.java
@@ -14,8 +14,6 @@
 */
 package io.github.mfaisalkhatri.tests.saucedemo;
 
-import io.github.mfaisalkhatri.pages.saucedemo.LoginPage;
-import io.github.mfaisalkhatri.pages.saucedemo.MainPage;
 import io.github.mfaisalkhatri.tests.base.BaseSuiteSetup;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
